requirements.txt
httpx~=0.28.1
allure-python-commons~=2.15.0
selenium~=4.35.0
pytest~=8.4.2
webdriver-manager~=4.0.2
pydantic~=2.11.9
pydantic-settings~=2.10.1
allure-pytest~=2.15.0
confluent-kafka~=2.11.1
jsonschema~=4.25.1
mimesis~=18.0.0